No longer set a nodata value
Previously I was setting NoData to 255, however this was causing entire rows to be missing from the output raster using numpy 1.13.1 and gdal 2.2.1 .  Setting to 0 fixes the missing rows, but there is no distinguishing between 0-value pixels and NoData.  This requires further investigation, but for now, simply not setting a NoData value seems to have mitigated the issue.